Removed 2d and 3d types

+++ Common.Meshing.Test/HalfEdgeBased/HBCreatePolygonMesh2Test.cs
@@ -13,8 +13,8 @@ namespace Common.Meshing.Test.HalfEdgeBased
         [TestMethod]
         public void FromBox()
         {
-            var min = new Vector2d(-1, -1);
-            var max = new Vector2d(1, 1);
+            var min = new Vector2f(-1, -1);
+            var max = new Vector2f(1, 1);
 
             var mesh = HBCreatePolygonMesh2.FromBox(min, max);
 
@@ -22,9 +22,9 @@ namespace Common.Meshing.Test.HalfEdgeBased
             Assert.AreEqual(8, mesh.Edges.Count);
             Assert.AreEqual(1, mesh.Faces.Count);
             Assert.AreEqual(min, mesh.Vertices[0].Position);
-            Assert.AreEqual(new Vector2d(max.x, min.y), mesh.Vertices[1].Position);
+            Assert.AreEqual(new Vector2f(max.x, min.y), mesh.Vertices[1].Position);
             Assert.AreEqual(max, mesh.Vertices[2].Position);
-            Assert.AreEqual(new Vector2d(min.x, max.y), mesh.Vertices[3].Position);
+            Assert.AreEqual(new Vector2f(min.x, max.y), mesh.Vertices[3].Position);
 
             HBMeshHelper.CheckVertex(mesh, vertex: 0, edge: 0);
             HBMeshHelper.CheckVertex(mesh, vertex: 1, edge: 1);
@@ -44,7 +44,7 @@ namespace Common.Meshing.Test.HalfEdgeBased
         [TestMethod]
         public void FromCircle()
         {
-            var mesh = HBCreatePolygonMesh2.FromCircle(Vector2d.Zero, 1.0, 4);
+            var mesh = HBCreatePolygonMesh2.FromCircle(Vector2f.Zero, 1.0f, 4);
 
             Assert.AreEqual(4, mesh.Vertices.Count);
             Assert.AreEqual(8, mesh.Edges.Count);
